+++ nametag3_server.py
@@ -90,7 +90,6 @@ import urllib.parse
 os.environ.setdefault("KERAS_BACKEND", "torch")
 
 import torch
-import transformers
 
 from nametag3_dataset import NameTag3Dataset
 from nametag3_dataset_collection import NameTag3DatasetCollection
@@ -163,8 +162,7 @@ class Models:
             print("Model options loaded successfully:\n{}".format(self._args), file=sys.stderr, flush=True)
 
             # Load the HF tokenizer
-            self.hf_tokenizer = transformers.AutoTokenizer.from_pretrained(self._args.hf_plm,
-                                                                           add_prefix_space = self._args.hf_plm in ["roberta-base", "roberta-large", "ufal/robeczech-base"])
+            self.hf_tokenizer = NameTag3Dataset.get_hf_tokenizer(self._args.hf_plm)
 
             # Unpickle word mappings of train data
             self._train_collection = NameTag3DatasetCollection(self._args,